+++ src/commands/CommandBlueprint.test.ts
@@ -1,14 +1,14 @@
 import { number, string } from "knub-command-manager";
 import { expect } from "chai";
-import { BasePluginType, globalCommand } from "..";
-import { guildCommand } from "./CommandBlueprint";
+import { BasePluginType } from "..";
+import { typedGuildCommand, typedGlobalCommand } from "./CommandBlueprint";
 import { GuildTextableChannel, PrivateChannel, Textable } from "eris";
 
 type AssertEquals<TActual, TExpected> = TActual extends TExpected ? true : false;
 
-describe("guildCommand() helper", () => {
+describe("typedGuildCommand() helper", () => {
   it("(blueprint)", () => {
-    const blueprint = guildCommand({
+    const blueprint = typedGuildCommand({
       trigger: "cmd",
       permission: null,
       signature: {
@@ -27,59 +27,6 @@ describe("guildCommand() helper", () => {
     expect(blueprint.run).to.not.equal(undefined);
   });
 
-  it("(trigger, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = guildCommand("cmd", () => {});
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.equal(undefined);
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
-  it("(trigger, signature, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = guildCommand(
-      "cmd",
-      {
-        foo: string(),
-        bar: number(),
-      },
-      ({ args }) => {
-        // Test type inference
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result: AssertEquals<typeof args, { foo: string; bar: number }> = true;
-      }
-    );
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.eql({ foo: string(), bar: number() });
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
-  it("(trigger, signature, options, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = guildCommand(
-      "cmd",
-      {
-        foo: string(),
-        bar: number(),
-      },
-      {
-        permission: "foo",
-      },
-      ({ args }) => {
-        // Test type inference
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result: AssertEquals<typeof args, { foo: string; bar: number }> = true;
-      }
-    );
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.eql({ foo: string(), bar: number() });
-    expect(blueprint.permission).to.equal("foo");
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
   interface CustomPluginType extends BasePluginType {
     state: {
       foo: 5;
@@ -87,7 +34,7 @@ describe("guildCommand() helper", () => {
   }
 
   it("<TPluginType>()(blueprint)", () => {
-    const blueprint = guildCommand<CustomPluginType>()({
+    const blueprint = typedGuildCommand<CustomPluginType>()({
       trigger: "cmd",
       permission: null,
       signature: {
@@ -108,89 +55,32 @@ describe("guildCommand() helper", () => {
     expect(blueprint.run).to.not.equal(undefined);
   });
 
-  it("<TPluginType>()(trigger, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = guildCommand<CustomPluginType>()("cmd", ({ pluginData }) => {
-      // Test type inference
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const result: AssertEquals<typeof pluginData.state.foo, 5> = true;
-    });
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.equal(undefined);
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
-  it("<TPluginType>()(trigger, signature, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = guildCommand<CustomPluginType>()(
-      "cmd",
-      {
-        foo: string(),
-        bar: number(),
-      },
-      ({ args, pluginData }) => {
-        // Test type inference
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result: AssertEquals<typeof args, { foo: string; bar: number }> = true;
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result2: AssertEquals<typeof pluginData.state.foo, 5> = true;
-      }
-    );
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.eql({ foo: string(), bar: number() });
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
-  it("<TPluginType>()(trigger, signature, options, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = guildCommand<CustomPluginType>()(
-      "cmd",
-      {
-        foo: string(),
-        bar: number(),
-      },
-      {
-        permission: "foo",
-      },
-      ({ args, pluginData }) => {
-        // Test type inference
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result: AssertEquals<typeof args, { foo: string; bar: number }> = true;
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result2: AssertEquals<typeof pluginData.state.foo, 5> = true;
-      }
-    );
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.eql({ foo: string(), bar: number() });
-    expect(blueprint.permission).to.equal("foo");
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
   it("command message is a guild message", () => {
-    guildCommand("foo", {}, ({ message }) => {
-      // Make sure message.member cannot be null
-      // https://github.com/microsoft/TypeScript/issues/29627#issuecomment-458329399
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const result: null extends typeof message.member ? false : true = true;
+    typedGuildCommand({
+      trigger: "foo",
+      permission: null,
+      run({ message }) {
+        // Make sure message.member cannot be null
+        // https://github.com/microsoft/TypeScript/issues/29627#issuecomment-458329399
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        const result: null extends typeof message.member ? false : true = true;
 
-      // Make sure message.channel is always a textable guild channel and cannot be a private channel
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const result2: Textable & PrivateChannel extends typeof message.channel
-        ? false
-        : GuildTextableChannel extends typeof message.channel
-        ? true
-        : false = true;
+        // Make sure message.channel is always a textable guild channel and cannot be a private channel
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        const result2: Textable & PrivateChannel extends typeof message.channel
+          ? false
+          : GuildTextableChannel extends typeof message.channel
+          ? true
+          : false = true;
+      },
     });
   });
 
   it("args type inference for multiple signatures", () => {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const blueprint = guildCommand(
-      "cmd",
-      [
+    typedGuildCommand({
+      trigger: "cmd",
+      permission: null,
+      signature: [
         {
           foo: string(),
           bar: number(),
@@ -199,7 +89,7 @@ describe("guildCommand() helper", () => {
           baz: number(),
         },
       ],
-      ({ args }) => {
+      run({ args }) {
         if (args.foo != null) {
           // eslint-disable-next-line @typescript-eslint/no-unused-vars
           const x: number = args.bar; // args.bar cannot be undefined
@@ -213,14 +103,14 @@ describe("guildCommand() helper", () => {
           // eslint-disable-next-line @typescript-eslint/no-unused-vars
           const y: undefined = args.bar; // args.bar must be undefined
         }
-      }
-    );
+      },
+    });
   });
 });
 
-describe("globalCommand() helper", () => {
+describe("typedGlobalCommand() helper", () => {
   it("(blueprint)", () => {
-    const blueprint = globalCommand({
+    const blueprint = typedGlobalCommand({
       trigger: "cmd",
       permission: null,
       signature: {
@@ -239,59 +129,6 @@ describe("globalCommand() helper", () => {
     expect(blueprint.run).to.not.equal(undefined);
   });
 
-  it("(trigger, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = globalCommand("cmd", () => {});
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.equal(undefined);
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
-  it("(trigger, signature, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = globalCommand(
-      "cmd",
-      {
-        foo: string(),
-        bar: number(),
-      },
-      ({ args }) => {
-        // Test type inference
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result: AssertEquals<typeof args, { foo: string; bar: number }> = true;
-      }
-    );
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.eql({ foo: string(), bar: number() });
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
-  it("(trigger, signature, options, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = globalCommand(
-      "cmd",
-      {
-        foo: string(),
-        bar: number(),
-      },
-      {
-        permission: "foo",
-      },
-      ({ args }) => {
-        // Test type inference
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result: AssertEquals<typeof args, { foo: string; bar: number }> = true;
-      }
-    );
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.eql({ foo: string(), bar: number() });
-    expect(blueprint.permission).to.equal("foo");
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
   interface CustomPluginType extends BasePluginType {
     state: {
       foo: 5;
@@ -299,7 +136,7 @@ describe("globalCommand() helper", () => {
   }
 
   it("<TPluginType>()(blueprint)", () => {
-    const blueprint = globalCommand<CustomPluginType>()({
+    const blueprint = typedGlobalCommand<CustomPluginType>()({
       trigger: "cmd",
       permission: null,
       signature: {
@@ -320,82 +157,25 @@ describe("globalCommand() helper", () => {
     expect(blueprint.run).to.not.equal(undefined);
   });
 
-  it("<TPluginType>()(trigger, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = globalCommand<CustomPluginType>()("cmd", ({ pluginData }) => {
-      // Test type inference
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const result: AssertEquals<typeof pluginData.state.foo, 5> = true;
-    });
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.equal(undefined);
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
-  it("<TPluginType>()(trigger, signature, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = globalCommand<CustomPluginType>()(
-      "cmd",
-      {
-        foo: string(),
-        bar: number(),
-      },
-      ({ args, pluginData }) => {
-        // Test type inference
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result: AssertEquals<typeof args, { foo: string; bar: number }> = true;
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result2: AssertEquals<typeof pluginData.state.foo, 5> = true;
-      }
-    );
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.eql({ foo: string(), bar: number() });
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
-  it("<TPluginType>()(trigger, signature, options, run)", () => {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    const blueprint = globalCommand<CustomPluginType>()(
-      "cmd",
-      {
-        foo: string(),
-        bar: number(),
-      },
-      {
-        permission: "foo",
-      },
-      ({ args, pluginData }) => {
-        // Test type inference
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result: AssertEquals<typeof args, { foo: string; bar: number }> = true;
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-        const result2: AssertEquals<typeof pluginData.state.foo, 5> = true;
-      }
-    );
-
-    expect(blueprint.trigger).to.equal("cmd");
-    expect(blueprint.signature).to.eql({ foo: string(), bar: number() });
-    expect(blueprint.permission).to.equal("foo");
-    expect(blueprint.run).to.not.equal(undefined);
-  });
-
   it("command message is NOT necessarily a guild message", () => {
-    globalCommand("foo", {}, ({ message }) => {
-      // If the message is not necessarily a guild message, the member can be null
-      // https://github.com/microsoft/TypeScript/issues/29627#issuecomment-458329399
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const result: null extends typeof message.member ? true : false = true;
+    typedGlobalCommand({
+      trigger: "foo",
+      permission: null,
+      run({ message }) {
+        // If the message is not necessarily a guild message, the member can be null
+        // https://github.com/microsoft/TypeScript/issues/29627#issuecomment-458329399
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        const result: null extends typeof message.member ? true : false = true;
 
-      // If the message is not necessarily a guild message, the channel can be a private channel
-      // as well as a guild channel.
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const result2: Textable & PrivateChannel extends typeof message.channel
-        ? GuildTextableChannel extends typeof message.channel
-          ? true
-          : false
-        : false = true;
+        // If the message is not necessarily a guild message, the channel can be a private channel
+        // as well as a guild channel.
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        const result2: Textable & PrivateChannel extends typeof message.channel
+          ? GuildTextableChannel extends typeof message.channel
+            ? true
+            : false
+          : false = true;
+      },
     });
   });
 });